Fix: Properly reload the deconstantized constant
Previously, the code was attempting to constantize an Array.

The goal is to remove the constant and reload it so it can pick
up changes, because it was reloading them, but having the old copy
around was messing things up: any class that had inherited from one
of these would keep the old copy as what it was inheriting from,
rather than the new one. So we wanted to remove the constant to
remove the reference, and add it back in to re-establish it.

The fix is to:
1) Remove all but the *last* constant in a namespaced klass name
2) Join the remainder back into a namespaced string
... and *then* constantize the result, as before.